Added theme configuration options (incomplete)
This is a basic POC implementation.
Needs a lot of cleanup and fixing.
