+++ Cargo.toml
@@ -35,6 +35,8 @@ structopt = "0.3.21"
 rand = "0.8.3"
 unicode-width = "0.1.8"
 serde = { version = "1.0.125", features = ["derive"] }
+dirs = "3.0.2"
+toml = "0.5.8"
 
 [profile.release]
 opt-level = 3
